Rename parser to live
