+++ src-tauri/src/live/party_tracker.rs
@@ -1,4 +1,4 @@
-use crate::parser::id_tracker::IdTracker;
+use crate::live::id_tracker::IdTracker;
 use hashbrown::{HashMap, HashSet};
 use std::cell::RefCell;
 use std::rc::Rc;
