+++ src-tauri/src/main.rs
@@ -4,7 +4,7 @@
 )]
 
 mod app;
-mod parser;
+mod live;
 
 use std::{
     fs::{self, File},
@@ -16,8 +16,8 @@ use std::{
 use anyhow::Result;
 use flate2::read::GzDecoder;
 use hashbrown::HashMap;
+use live::models::*;
 use log::{error, info, warn};
-use parser::models::*;
 
 use rusqlite::{params, params_from_iter, Connection, Transaction};
 use sysinfo::System;
@@ -69,7 +69,7 @@ async fn main() -> Result<()> {
         .add_item(quit);
 
     let system_tray = SystemTray::new().with_menu(tray_menu);
-    
+
     tauri::Builder::default()
         .setup(|app| {
             info!("starting app v{}", app.package_info().version.to_string());
@@ -165,7 +165,7 @@ async fn main() -> Result<()> {
             info!("listening on port: {}", port);
             remove_driver();
             task::spawn_blocking(move || {
-                parser::start(meter_window, port, settings).map_err(|e| {
+                live::start(meter_window, port, settings).map_err(|e| {
                     error!("unexpected error occurred in parser: {}", e);
                 })
             });
@@ -1519,7 +1519,10 @@ fn set_clickthrough(window: tauri::Window, set: bool) {
 
 #[tauri::command]
 fn remove_driver() {
-    Command::new("sc").args(["delete", "windivert"]).output().expect("unable to delete driver");
+    Command::new("sc")
+        .args(["delete", "windivert"])
+        .output()
+        .expect("unable to delete driver");
 }
 
 #[tauri::command]
@@ -1566,11 +1569,20 @@ fn set_start_on_boot(set: bool) {
     if set {
         Command::new("schtasks")
             .args(["/delete", "/tn", task_name, "/f"])
-            .output().ok();
-        
+            .output()
+            .ok();
+
         let output = Command::new("schtasks")
             .args([
-                "/create", "/tn", task_name, "/tr", &format!("\"{}\"", &app_path), "/sc", "onlogon", "/rl", "highest",
+                "/create",
+                "/tn",
+                task_name,
+                "/tr",
+                &format!("\"{}\"", &app_path),
+                "/sc",
+                "onlogon",
+                "/rl",
+                "highest",
             ])
             .output();
